Fix: s3 이미지 삭제하는 부분에 default 값일 경우 삭제하지 않는 부분 추가되었습니다.
Resolves: #16

+++ src/main/java/com/zerobase/foodlier/global/profile/facade/ProfileFacade.java
@@ -1,6 +1,7 @@
 package com.zerobase.foodlier.global.profile.facade;
 
 import com.zerobase.foodlier.common.s3.service.S3Service;
+import com.zerobase.foodlier.module.member.member.constants.ProfileUrlConstants;
 import com.zerobase.foodlier.module.member.member.profile.dto.MemberPrivateProfileForm;
 import com.zerobase.foodlier.module.member.member.domain.model.Member;
 import com.zerobase.foodlier.module.member.member.local.dto.CoordinateResponseDto;
@@ -29,7 +30,9 @@ public class ProfileFacade {
         String imageUrl = form.getProfileImage() != null ?
                 s3Service.getImageUrl(form.getProfileImage()) :
                 member.getProfileUrl();
-        s3Service.deleteImage(member.getProfileUrl());
+        if (!imageUrl.equals(ProfileUrlConstants.PROFILE_DEFAULT_URL)) {
+            s3Service.deleteImage(member.getProfileUrl());
+        }
         memberService.updatePrivateProfile(
                 MemberUpdateDto.from(
                         form, coordinateResponseDto, imageUrl
